+++ basic_bootloader/src/bootloader/process_transaction.rs
@@ -616,6 +616,7 @@ where
         };
 
         // Parse, validate and apply authorization list, following EIP-7702
+        #[cfg(feature = "pectra")]
         transaction.parse_authorization_list_and_apply_delegations(system, &mut resources)?;
 
         // Take a snapshot in case we need to revert due to out of native.
