+++ basic_bootloader/src/bootloader/transaction/mod.rs
@@ -1267,6 +1267,7 @@ impl<'a> ZkSyncTransaction<'a> {
     ///
     /// Parse length of authorization list.
     ///
+    #[cfg(feature = "pectra")]
     pub fn parse_authorization_list_length(&self) -> Result<u64, TxError> {
         let iter = self
             .reserved_dynamic
@@ -1278,6 +1279,7 @@ impl<'a> ZkSyncTransaction<'a> {
     ///
     /// Parse and validate authorization list, while applying delegations.
     ///
+    #[cfg(feature = "pectra")]
     pub fn parse_authorization_list_and_apply_delegations<S: EthereumLikeTypes>(
         &self,
         system: &mut System<S>,
@@ -1513,4 +1515,20 @@ fn estimate_authorization_list_item_length(item: &AuthorizationListItem) -> usiz
         + rlp::estimate_number_encoding_len(&y_parity.to_be_bytes())
         + rlp::estimate_number_encoding_len(&r.to_be_bytes::<32>())
         + rlp::estimate_number_encoding_len(&s.to_be_bytes::<32>())
+}
+
+///
+/// Parse a delegation of the format: 0xef0100 || address
+///
+pub fn parse_delegation(delegation: &[u8]) -> Result<B160, InternalError> {
+    if delegation.len() != EIP7702_DELEGATION_MARKER.len() + B160::BYTES {
+        return Err(internal_error!("7702 delegation of incorrect length"));
+    }
+    if delegation[0..3] != EIP7702_DELEGATION_MARKER {
+        return Err(internal_error!("7702 delegation has invalid prefix"));
+    }
+    let Some(address) = B160::try_from_be_slice(&delegation[3..]) else {
+        return Err(internal_error!("7702 delegation has invalid address"));
+    };
+    Ok(address)
 }
